requirements.txt
asgiref==3.8.1
certifi==2024.6.2
distlib==0.3.8
Django==4.2.13
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
Faker==25.8.0
filelock==3.15.1
pipenv==2024.0.1
platformdirs==4.2.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
virtualenv==20.26.2
